+++ client/rpc/src/main/kotlin/net/corda/client/rpc/CordaRPCClient.kt
@@ -11,7 +11,8 @@ import net.corda.core.messaging.ClientRpcSslOptions
 import net.corda.core.utilities.days
 import net.corda.core.utilities.minutes
 import net.corda.core.utilities.seconds
-import net.corda.nodeapi.internal.InternalArtemisTcpTransport.Companion.rpcConnectorTcpTransport
+import net.corda.nodeapi.internal.ArtemisTcpTransport.Companion.rpcConnectorTcpTransport
+import net.corda.nodeapi.internal.PLATFORM_VERSION
 import net.corda.serialization.internal.AMQP_RPC_CLIENT_CONTEXT
 import java.time.Duration
 
@@ -45,7 +46,7 @@ open class CordaRPCClientConfiguration @JvmOverloads constructor(
          * The default value is whatever version of Corda this RPC library was shipped as a part of. Therefore if you
          * use the RPC library from Corda 4, it will by default only connect to a node of version 4 or above.
          */
-        open val minimumServerProtocolVersion: Int = 4,
+        open val minimumServerProtocolVersion: Int = PLATFORM_VERSION,
 
         /**
          * If set to true the client will track RPC call sites (default is false). If an error occurs subsequently