+++ node-api/src/main/kotlin/net/corda/nodeapi/internal/NodeInfoConstants.kt
@@ -2,3 +2,4 @@ package net.corda.nodeapi.internal
 
 // TODO: Add to Corda node.conf to allow customisation
 const val NODE_INFO_DIRECTORY = "additional-node-infos"
+const val PLATFORM_VERSION = 4