+++ node/src/main/kotlin/net/corda/node/services/messaging/P2PMessagingClient.kt
@@ -33,7 +33,7 @@ import net.corda.nodeapi.internal.ArtemisMessagingComponent.Companion.BRIDGE_NOT
 import net.corda.nodeapi.internal.ArtemisMessagingComponent.Companion.JOURNAL_HEADER_SIZE
 import net.corda.nodeapi.internal.ArtemisMessagingComponent.Companion.P2PMessagingHeaders
 import net.corda.nodeapi.internal.ArtemisMessagingComponent.Companion.PEERS_PREFIX
-import net.corda.nodeapi.internal.InternalArtemisTcpTransport.Companion.p2pConnectorTcpTransport
+import net.corda.nodeapi.internal.ArtemisTcpTransport.Companion.p2pConnectorTcpTransport
 import net.corda.nodeapi.internal.bridging.BridgeControl
 import net.corda.nodeapi.internal.bridging.BridgeEntry
 import net.corda.nodeapi.internal.persistence.CordaPersistence
@@ -120,7 +120,7 @@ class P2PMessagingClient(val config: NodeConfiguration,
     private lateinit var advertisedAddress: NetworkHostAndPort
     private var maxMessageSize: Int = -1
 
-    override val myAddress: SingleMessageRecipient get() = NodeAddress(myIdentity, advertisedAddress)
+    override val myAddress: SingleMessageRecipient get() = NodeAddress(myIdentity)
     override val ourSenderUUID = UUID.randomUUID().toString()
 
     private val state = ThreadBox(InnerState())
@@ -233,7 +233,7 @@ class P2PMessagingClient(val config: NodeConfiguration,
         fun gatherAddresses(node: NodeInfo): Sequence<BridgeEntry> {
             return state.locked {
                 node.legalIdentitiesAndCerts.map {
-                    val messagingAddress = NodeAddress(it.party.owningKey, node.addresses.first())
+                    val messagingAddress = NodeAddress(it.party.owningKey)
                     BridgeEntry(messagingAddress.queueName, node.addresses, node.legalIdentities.map { it.name })
                 }.filter { producerSession!!.queueQuery(SimpleString(it.queueName)).isExists }.asSequence()
             }
@@ -242,14 +242,14 @@ class P2PMessagingClient(val config: NodeConfiguration,
         fun deployBridges(node: NodeInfo) {
             gatherAddresses(node)
                     .forEach {
-                        sendBridgeControl(BridgeControl.Create(myIdentity.toStringShort(), it))
+                        sendBridgeControl(BridgeControl.Create(config.myLegalName.toString(), it))
                     }
         }
 
         fun destroyBridges(node: NodeInfo) {
             gatherAddresses(node)
                     .forEach {
-                        sendBridgeControl(BridgeControl.Delete(myIdentity.toStringShort(), it))
+                        sendBridgeControl(BridgeControl.Delete(config.myLegalName.toString(), it))
                     }
         }
 
@@ -289,7 +289,7 @@ class P2PMessagingClient(val config: NodeConfiguration,
                 delayStartQueues += queue.toString()
             }
         }
-        val startupMessage = BridgeControl.NodeToBridgeSnapshot(myIdentity.toStringShort(), inboxes, requiredBridges)
+        val startupMessage = BridgeControl.NodeToBridgeSnapshot(config.myLegalName.toString(), inboxes, requiredBridges)
         sendBridgeControl(startupMessage)
     }
 
@@ -310,12 +310,9 @@ class P2PMessagingClient(val config: NodeConfiguration,
                     return
                 }
                 eventsSubscription = p2pConsumer!!.messages
-                        .doOnError { error -> throw error }
-                        .doOnNext { message -> deliver(message) }
                         // this `run()` method is semantically meant to block until the message consumption runs, hence the latch here
                         .doOnCompleted(latch::countDown)
-                        .doOnError { error -> throw error }
-                        .subscribe()
+                        .subscribe({ message -> deliver(message) }, { error -> throw error })
                 p2pConsumer!!
             }
             consumer.start()
@@ -495,7 +492,7 @@ class P2PMessagingClient(val config: NodeConfiguration,
             val peers = networkMap.getNodesByOwningKeyIndex(keyHash)
             for (node in peers) {
                 val bridge = BridgeEntry(queueName, node.addresses, node.legalIdentities.map { it.name })
-                val createBridgeMessage = BridgeControl.Create(myIdentity.toStringShort(), bridge)
+                val createBridgeMessage = BridgeControl.Create(config.myLegalName.toString(), bridge)
                 sendBridgeControl(createBridgeMessage)
             }
         }
@@ -540,7 +537,7 @@ class P2PMessagingClient(val config: NodeConfiguration,
 
     override fun getAddressOfParty(partyInfo: PartyInfo): MessageRecipients {
         return when (partyInfo) {
-            is PartyInfo.SingleNode -> NodeAddress(partyInfo.party.owningKey, partyInfo.addresses.single())
+            is PartyInfo.SingleNode -> NodeAddress(partyInfo.party.owningKey)
             is PartyInfo.DistributedNode -> ServiceAddress(partyInfo.party.owningKey)
         }
     }