+++ docs/source/changelog.rst
@@ -9,6 +9,8 @@ Unreleased
 
 * Case insensitive vault queries can be specified via a boolean on applicable SQL criteria builder operators. By default queries will be case sensitive.
 
+* Getter added to ``CordaRPCOps`` for the node's network parameters.
+
 * The RPC client library now checks at startup whether the server is of the client libraries major version or higher.
   Therefore to connect to a Corda 4 node you must use version 4 or lower of the library. This behaviour can be overridden
   by specifying a lower number in the ``CordaRPCClientConfiguration`` class.